+++ app/models/volumes_from.rb
@@ -1,0 +1,7 @@
+class VolumesFrom < BaseResource
+
+  schema do
+    integer :service_id
+    string :service_name
+  end
+end