+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 Latest
 ------------------
 ### Fixed
-- don't show scrollbar on envirnment variable field (#347)
+- Don't show scrollbar on environment variable field (#347)
+
+### Added
+- Edit Service Name
+- Remove Multiple Images
 
 0.2.0 - 2014-09-09
 ------------------