+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-Latest
+0.2.4 - 2014-10-09
+------------------
+
+### Added
+- GitHub token update (#375)
+
+
+0.2.3 - 2014-10-02
 ------------------
 
 ### Fixed
@@ -16,7 +23,6 @@ Latest
  - Table-based layout for ports information (#369)
  - Show port numbers automatically assigned by Docker (#376)
  - Disable endpoint links when service is not running (#378, #379)
-- GitHub token update (#375)
 
 
 0.2.2 - 2014-09-26